refactor: silence lint warnings
The following two functions had lint warning from 'package-lint'

(citar-denote-keywords-prompt): Variable 'denote-last-keywords' is not
defined in denote or anywhere in denote-citar; it does not seem to be
used.  I thought we could remove it completely?

(citar-denote-has-notes): prefer 'let' to 'setq' for this?  The lambda
function seems to work inside the let scoping (otherwise, citar would
have complained about it).