refactor: add-reference to look for blank line
'citar-denote-add-reference' now has fall-back logic to look for a blank
line when the keywords property is not present in the front matter (the
original logic with 'char-after').

Denote allows user configuration to not have keywords property in the
front matter.  citar-denote also allows for regexp to search for
bibliographic notes other than "_bib" -- e.g. it can be a sub-directory
omitting keywords entirely.