Avoid duplicate citar-denote-keyword in citar-denote-keywords-prompt
I just started maintaining my own BibTeX bibliography and will be
using citar-denote with it.  While creating a new note with
citar-denote, I typed in the "bib" keyword, thinking that I had to do
this manually.  The result was a file name with duplicate "bib"
keywords.

This change addresses that issue.  It also uses the standard
'denote-keywords-sort' function instead of doing its own check.  In
practice, this is the same though it is easier to maintain
citar-denote in case we ever make changes to denote-keywords-sort.