update citar-denote.org to ensure citar-denote-open-note is callable
Bind `citar-denote-open-note` in `:preface` to ensure it is callable even when
citar-denote is not loaded. Then Emacs will load denote, citar and citar-denote
in sequence.

Notice the `:demand t` is also required. Otherwise, when citar or denote is
loaded, citar-denote won't be loaded automatically because any package that has
`:bind` or `:hook` is automatically set as `:defer t`, you need to call its
functions to load it.