introduce citar-denote-cite-includes-reference option
By default, citar-denote distinguishes references from citations. This
means `citar-denote--retrieve-cite-files' does not return files
that only contain the citekey in the frontmatter and do not cite the
bibliographic entry in the text.

This commit introduces a custom variable to change this behaviour.
When `citar-denote-cite-includes-reference' is non-nil, retrieving
cited files will include files that only contain the citekey in the
front matter.

The effect is achieved by not including the '@' sign in the xref
query (when the option is non-nil).